actual calculaton of flags

+++ server/helpers/database.js
@@ -5491,9 +5491,17 @@ const database = {
     },
     addInstanceStaff: async (user_id, privilege) => {
         try {
+            const user = await database.getAccountByUserId(user_id);
+
+            if (user == null) {
+                logText(error, "User not found.");
+
+                return false;
+            }
+
             await database.runQuery(`INSERT INTO staff (user_id, privilege, audit_log) VALUES ($1, $2, $3)`, [user_id, privilege, '[]']);
 
-            await database.runQuery(`UPDATE users SET flags = 1 WHERE id = $1`, [user_id]); //todo calculate the flags and add onto existing
+            await database.runQuery(`UPDATE users SET flags = $1 WHERE id = $2`, [user.flags + 1, user_id]); // Maybe repurpose COLLABORATOR or RESTRICTED_COLLABORATOR for either Admin and Janitor/Moderator respectively
 
             return true;
         }  catch (error) {
@@ -5534,9 +5542,17 @@ const database = {
     },
     removeFromStaff: async (user_id) => {
         try {
+            const user = await database.getAccountByUserId(user_id);
+
+            if (user == null) {
+                logText(error, "User not found.");
+
+                return false;
+            }
+
             await database.runQuery(`DELETE FROM staff WHERE user_id = $1`, [user_id]);
 
-            await database.runQuery(`UPDATE users SET flags = 0 WHERE id = $1`, [user_id]);
+            await database.runQuery(`UPDATE users SET flags = $1 WHERE id = $2`, [user.flags - 1, user_id]); // Maybe repurpose COLLABORATOR or RESTRICTED_COLLABORATOR for either Admin and Janitor/Moderator respectively
 
             return true;
         } catch (error) {
